requirements.txt
vanna==0.7.9
streamlit==1.50.0
python-dotenv==1.0.1
psycopg2-binary==2.9.11